# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Firebase authentication service with social login support.",
        "title": "Auth Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/home": {
            "get": {
                "description": "Get personalized home page data including recommended stores and pickup times",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "home"
                ],
                "summary": "Get home page data",
                "parameters": [
                    {
                        "type": "number",
                        "description": "User's latitude",
                        "name": "latitude",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "User's longitude",
                        "name": "longitude",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.HomePageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/home/search": {
            "get": {
                "description": "Search for stores by name or description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "home"
                ],
                "summary": "Search stores",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "query",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.Store"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.HomePageResponse": {
            "type": "object",
            "properties": {
                "emailVerified": {
                    "type": "boolean"
                },
                "pickUpTomorrow": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.Store"
                    }
                },
                "recommendedStores": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.Store"
                    }
                },
                "userLocation": {
                    "type": "object",
                    "properties": {
                        "city": {
                            "type": "string"
                        },
                        "distance": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
...
        "handlers.Store": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "distance": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "isSaved": {
                    "type": "boolean"
                },
                "pickUpTime": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "rating": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}